Move simulated tests into receiver_simulated_test.go so we can keep track of which ones we need to port over.
